Proyecto/Proyecto/requirements.txt
django-crispy-forms
Django>=5.0
crispy-bootstrap4
Pillow
